Use u32 instead of usize for PID

+++ binaries/daemon/src/spawn.rs
@@ -42,7 +42,7 @@ pub async fn spawn_node(
     daemon_tx: mpsc::Sender<Timestamped<Event>>,
     dataflow_descriptor: Descriptor,
     clock: Arc<HLC>,
-) -> eyre::Result<usize> {
+) -> eyre::Result<u32> {
     let node_id = node.id.clone();
     tracing::debug!("Spawning node `{dataflow_id}/{node_id}`");
 
@@ -264,7 +264,7 @@ pub async fn spawn_node(
         .expect("Failed to create log file");
     let mut child_stdout =
         tokio::io::BufReader::new(child.stdout.take().expect("failed to take stdout"));
-    let pid = child.id().unwrap() as usize; // Todo: Manage failure
+    let pid = child.id().unwrap();
     let stdout_tx = tx.clone();
 
     // Stdout listener stream
